Task #1, ajuste de código modularizado ajustados arquivos para modularizar o código, aumentando a legibilidade e escalabilidade
