+++ .gitignore
@@ -1,2 +1,3 @@
 /Versões/
 .env
+__pycache__
